Initial commit
Just a starting point. Do something creative with this. It could be kind of fun to make it a long series of things you do to manipulate and change the DOM.

(Future self: You turned a one-hour, early-week assignment into a long (but enjoyable!) walk through a dark forest. You kind of got lost...)